Move `create`, `list` operations to separate packages, and rename their main functions to `Run`
