--- pkg/lib/operation/project/remote/workspace/operation.go
@@ -1,1 +1,0 @@
-package workspace
